requirements.txt
numpy==1.25.2
opencv_contrib_python==4.8.0.76
opencv_python==4.8.0.76
opencv_python_headless==4.10.0.84
pandas==2.0.3
Pillow==9.4.0
streamlit==1.36.0
streamlit_drawable_canvas==0.9.3
tensorflow==2.15.0
